md5_hunter/requirements_md5_hunter.txt
pdfplumber
python-docx